+++ gluten-core/src/main/scala/org/apache/gluten/extension/columnar/offload/OffloadSingleNode.scala
@@ -16,10 +16,17 @@
  */
 package org.apache.gluten.extension.columnar.offload
 
-import org.apache.gluten.extension.columnar.SingleNodeOps._
+import org.apache.gluten.execution.GlutenPlan
+import org.apache.gluten.extension.columnar.transition.Convention
 
+import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.execution.SparkPlan
+import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.catalyst.expressions.{Attribute, SortOrder}
+import org.apache.spark.sql.catalyst.plans.physical.Partitioning
+import org.apache.spark.sql.execution.{LeafExecNode, SparkPlan}
+import org.apache.spark.sql.vectorized.ColumnarBatch
 
 /**
  * Converts a vanilla Spark plan node into Gluten plan node. Gluten plan is supposed to be executed
@@ -29,6 +36,73 @@ import org.apache.spark.sql.execution.SparkPlan
  * the children node. Tree-walking is done by caller of this trait.
  */
 trait OffloadSingleNode extends Logging {
-  final def offload(plan: SparkPlan): SparkPlan = plan.applyOnNode(offload0)
-  protected[OffloadSingleNode] def offload0(plan: SparkPlan): SparkPlan
+  def offload(plan: SparkPlan): SparkPlan
+}
+
+object OffloadSingleNode {
+  implicit class OffloadSingleNodeOps(rule: OffloadSingleNode) {
+
+    /**
+     * Converts the [[OffloadSingleNode]] rule to strict version.
+     *
+     * In the strict version rule, all children of the input query plan node will be replaced with
+     * 'DummyLeafExec' nodes so they are not accessible in the rule body.
+     */
+    def toStrcitRule(): OffloadSingleNode = {
+      new StrictRule(rule);
+    }
+  }
+
+  private class StrictRule(delegate: OffloadSingleNode) extends OffloadSingleNode {
+    override def offload(plan: SparkPlan): SparkPlan = {
+      val planWithChildrenHidden = hideChildren(plan)
+      val applied = delegate.offload(planWithChildrenHidden)
+      val out = restoreHiddenChildren(applied)
+      out
+    }
+
+    /**
+     * Replaces the children with 'DummyLeafExec' nodes so they become inaccessible afterward. Used
+     * when the children plan nodes can be dropped because not interested.
+     */
+    private def hideChildren[T <: SparkPlan](plan: T): T = {
+      plan.withNewChildren(plan.children.map(child => new DummyLeafExec(child))).asInstanceOf[T]
+    }
+
+    /**
+     * Restores hidden children from the replaced 'DummyLeafExec' nodes. It's exposed only for
+     * compatibility reason. Not recommended to be used in formal rule code.
+     */
+    private def restoreHiddenChildren[T <: SparkPlan](plan: T): T = {
+      plan
+        .transformDown {
+          case d: DummyLeafExec =>
+            d.hiddenPlan
+          case other => other
+        }
+        .asInstanceOf[T]
+    }
+  }
+
+  /**
+   * The plan node that hides the real child plan node during #applyOnNode call. This is used when
+   * query planner doesn't allow a rule to access the child plan nodes from the input query plan
+   * node.
+   */
+  private case class DummyLeafExec(hiddenPlan: SparkPlan) extends LeafExecNode with GlutenPlan {
+    private lazy val conv: Convention = Convention.get(hiddenPlan)
+
+    override def batchType(): Convention.BatchType = conv.batchType
+    override def rowType0(): Convention.RowType = conv.rowType
+    override def output: Seq[Attribute] = hiddenPlan.output
+    override def outputPartitioning: Partitioning = hiddenPlan.outputPartitioning
+    override def outputOrdering: Seq[SortOrder] = hiddenPlan.outputOrdering
+
+    override def doExecute(): RDD[InternalRow] =
+      throw new UnsupportedOperationException("Not allowed in #applyOnNode call")
+    override def doExecuteColumnar(): RDD[ColumnarBatch] =
+      throw new UnsupportedOperationException("Not allowed in #applyOnNode call")
+    override def doExecuteBroadcast[T](): Broadcast[T] =
+      throw new UnsupportedOperationException("Not allowed in #applyOnNode call")
+  }
 }
