+++ gluten-iceberg/src/main/scala/org/apache/gluten/execution/OffloadIcebergScan.scala
@@ -26,7 +26,7 @@ import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.execution.datasources.v2.BatchScanExec
 
 case class OffloadIcebergScan() extends OffloadSingleNode {
-  override def offload0(plan: SparkPlan): SparkPlan = plan match {
+  override def offload(plan: SparkPlan): SparkPlan = plan match {
     case scan: BatchScanExec if IcebergScanTransformer.supportsBatchScan(scan.scan) =>
       IcebergScanTransformer(scan)
     case other => other
